Fix(?) variable declarations
Although the grammar specified in the articles uses the VAR keyword once
to begin multiple variable declarations, at least some of the sample
programs in the repository use VAR to begin each variable declaration.
This modifies the grammar in our implementation to adhere to the latter
case, making the former case no longer valid.